[BChain] Basic PoS validator selection
For each slot in an epoch a validator is pseudo-randomly selected.
The RNG is seeded with last block hash of previous epoch and a slot
number, so that it can be deterministically derived by all Users in
P2P network.

Current apporach shortcommings:
One problem is caused by 1:1 mapping between block slot and a
validator:
- With larger epoch size that might not be feasible to calculate
a validator for each slot and store the hash.
- it does not result it fair proportionallity.
- with greater number of stakers than slots, some will never get
picked
- with stakers less than slots, the pool will be prematurely
exhausted

Other problems:
- stake_pool can be modified during current epoch, i.e. when seed
of RNG is known
- stake_pool can differ from each peer perspective, i.e. there is no
stake_pool consensus mechanism
- the validator_to_slot_assignment vector is fully calculated for
repetitevely for each slot.