[BChain] Validators mod extraction
The concept of a Validator(s) in PoS-based blockchain seems important
enough to be wrapped in dedicated module. At this point minimal
coupling with other domain-driven componets is made thus all logic
related both Validatos and Wallets or Blocks is still kept at lib.rs.

The current implementaion of the validators contains (not limited to)
following problems:
- stake_pool is explict String-key HashMap
- Validators for given epoch are kept as Vec<String> althugh the
length of the container cannot be changed (at the moment)
- current_epoch validators are return so that the vec can be iterated,
maybe some interator implementation or use(Fn(Validator)) could
improve the design
- update_validators repeatedly perform seed.clone() for all validators