[BChain] total_stake removed
The total_stake was not necessary as user_reward calculations was
essentially cancelling it out

Additionaly discovered a bug:
- Guaranteed max reward exploit: The distribute_rewards i calculated
based on all stakes, even made post validator selection. Thus a
Validator can maxout reward on staking all it has in the current epoch
and withdraw it on the first block of the next epoch. The
distirbute_reward must take into account stake-pool from up to the
Validator selection epoch.